+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 11 February 2022
+
+##### version 1.0.1
+
+Improved architecture of Resource Controller, init method was moved to template method of new entity called templates.
+I also created a core folder, I plan to store there some design features.
+
 ## 08 February 2022
 
 ##### version 0.4.3-beta